Added NSC Pointer chasing. Refactored code, moved computation helper functions to another file
