+++ hw/block/femu/femu.c
@@ -12,6 +12,8 @@
 
 #include <unistd.h>
 #include <sys/syscall.h>
+#include "computation.h"
+#include "computation.c"
 
 extern uint64_t iscos_counter;
 
@@ -122,32 +124,6 @@ static void nvme_process_cq_cpl(void *arg, int index_poller)
             break;
     }
 }
-
-int count_bits(char *buf)
-{
-        int i, j;
-        char c;
-        int count =0;
-
-        for (i = 0 ; i < 4096 ; i++) {
-                c = buf[i];
-                for (j = 0 ; j < 8 ; j++) {
-                        if (c & (1 << j)) {
-                                count +=  1;
-                        }
-                }
-        }
-        return count;
-}
-
-int get_disk_pointer(char *buf)
-{
-	int *p = buf;
-	return p[0];
-}
-
-#define COUNTING
-#define POINTER_CHASING
 
 void computational_thread ()
 {
